@@ -14,7 +14,7 @@ class ImageConversions {
 
     int h = rgb.getHeight(shape);
     int w = rgb.getWidth(shape);
-    Image image = Image(w, h);
+    Image image = Image(width: w, height: h);
 
     List<int> rgbValues = buffer.getIntList();
     assert(rgbValues.length == w * h * 3);
@@ -23,7 +23,7 @@ class ImageConversions {
       int r = rgbValues[j++];
       int g = rgbValues[j++];
       int b = rgbValues[j++];
-      image.setPixelRgba(wi, hi, r, g, b);
+      image.setPixelRgb(wi, hi, r, g, b);
       wi++;
       if (wi % w == 0) {
         wi = 0;
@@ -36,17 +36,16 @@ class ImageConversions {
 
   static Image convertGrayscaleTensorBufferToImage(TensorBuffer buffer) {
     // Convert buffer into Uint8 as needed.
-    TensorBuffer uint8Buffer = buffer.getDataType() == TfLiteType.uint8
+    TensorBuffer uint8Buffer = buffer.getDataType() == TfLiteType.kTfLiteUInt8
         ? buffer
-        : TensorBuffer.createFrom(buffer, TfLiteType.uint8);
+        : TensorBuffer.createFrom(buffer, TfLiteType.kTfLiteUInt8);
 
     final shape = uint8Buffer.getShape();
     final grayscale = ColorSpaceType.GRAYSCALE;
     grayscale.assertShape(shape);
 
-    final image = Image.fromBytes(grayscale.getWidth(shape),
-        grayscale.getHeight(shape), uint8Buffer.getBuffer().asUint8List(),
-        format: Format.luminance);
+    final image = Image.fromBytes(width: grayscale.getWidth(shape),
+        height: grayscale.getHeight(shape), bytes: uint8Buffer.getBuffer());
 
     return image;
   }
@@ -54,11 +53,11 @@ class ImageConversions {
   static void convertImageToTensorBuffer(Image image, TensorBuffer buffer) {
     int w = image.width;
     int h = image.height;
-    List<int> intValues = image.data;
+    List<int> intValues = image.toUint8List();
     int flatSize = w * h * 3;
     List<int> shape = [h, w, 3];
     switch (buffer.getDataType()) {
-      case TfLiteType.uint8:
+      case TfLiteType.kTfLiteUInt8:
         List<int> byteArr = List.filled(flatSize, 0);
         for (int i = 0, j = 0; i < intValues.length; i++) {
           byteArr[j++] = ((intValues[i]) & 0xFF);
@@ -67,7 +66,7 @@ class ImageConversions {
         }
         buffer.loadList(byteArr, shape: shape);
         break;
-      case TfLiteType.float32:
+      case TfLiteType.kTfLiteFloat32:
         List<double> floatArr = List.filled(flatSize, 0.0);
         for (int i = 0, j = 0; i < intValues.length; i++) {
           floatArr[j++] = ((intValues[i]) & 0xFF).toDouble();
