@@ -18,7 +18,7 @@ class Rot90Op extends ImageOperator {
   /// with the output.
   @override
   TensorImage apply(TensorImage image) {
-    Image rotated = copyRotate(image.image, 90 * _numRotation);
+    Image rotated = copyRotate(image.image, angle: 90 * _numRotation);
     image.loadImage(rotated);
     return image;
   }
