@@ -58,7 +58,7 @@ class TensorBufferContainer implements BaseImageContainer {
 
   @override
   Image get image {
-    if (_buffer.getDataType() != TfLiteType.uint8) {
+    if (_buffer.getDataType() != TfLiteType.kTfLiteUInt8) {
       // Print warning instead of throwing an exception. When using float models, users may want to
       // convert the resulting float image into Bitmap. That's fine to do so, as long as they are
       // aware of the potential accuracy lost when casting to uint8.
@@ -72,7 +72,7 @@ class TensorBufferContainer implements BaseImageContainer {
   }
 
   @override
-  TensorBuffer getTensorBuffer(TfLiteType dataType) {
+  TensorBuffer getTensorBuffer(int dataType) {
     // If the data type of buffer is desired, return it directly. Not making a defensive copy for
     // performance considerations. During image processing, users may need to set and get the
     // TensorBuffer many times.
