@@ -30,30 +30,30 @@ abstract class TensorBuffer {
   /// ```dart
   /// Creating a float TensorBuffer with shape [2, 3]:
   /// List<int> shape = [2, 3];
-  /// TensorBuffer tensorBuffer = TensorBuffer.createFixedSize(shape, TfLiteType.float32);
+  /// TensorBuffer tensorBuffer = TensorBuffer.createFixedSize(shape, TfLiteType.kTfLiteFloat32);
   /// ```
   ///
   /// ```dart
   /// Creating an uint8 TensorBuffer of a scalar:
   /// List<int> shape = [2, 3];
-  /// TensorBuffer tensorBuffer = TensorBuffer.createFixedSize(shape, TfLiteType.uint8);
+  /// TensorBuffer tensorBuffer = TensorBuffer.createFixedSize(shape, TfLiteType.kTfLiteUInt8);
   /// ```
   ///
   /// ```dart
   /// Creating an empty uint8 TensorBuffer:
   /// List<int> shape = [0];
-  /// TensorBuffer tensorBuffer = TensorBuffer.createFixedSize(shape, TfLiteType.uint8);
+  /// TensorBuffer tensorBuffer = TensorBuffer.createFixedSize(shape, TfLiteType.kTfLiteUInt8);
   /// ```
   ///
   /// The size of a fixed-size TensorBuffer cannot be changed once it is created.
   ///
   /// Throws [ArgumentError.notNull] if [shape] is null and
   /// [ArgumentError] is [shape] has non-positive elements.
-  static TensorBuffer createFixedSize(List<int> shape, TfLiteType dataType) {
+  static TensorBuffer createFixedSize(List<int> shape, int dataType) {
     switch (dataType) {
-      case TfLiteType.float32:
+      case TfLiteType.kTfLiteFloat32:
         return TensorBufferFloat(shape);
-      case TfLiteType.uint8:
+      case TfLiteType.kTfLiteUInt8:
         return TensorBufferUint8(shape);
       default:
         throw ArgumentError(
@@ -68,11 +68,11 @@ abstract class TensorBuffer {
   ///
   /// Dynamic TensorBuffers will reallocate memory when loading arrays or data buffers of
   /// different buffer sizes.
-  static TensorBuffer createDynamic(TfLiteType dataType) {
+  static TensorBuffer createDynamic(int dataType) {
     switch (dataType) {
-      case TfLiteType.float32:
+      case TfLiteType.kTfLiteFloat32:
         return TensorBufferFloat.dynamic();
-      case TfLiteType.uint8:
+      case TfLiteType.kTfLiteUInt8:
         return TensorBufferUint8.dynamic();
       default:
         throw ArgumentError(
@@ -83,7 +83,7 @@ abstract class TensorBuffer {
   /// Creates a [TensorBuffer] deep-copying data from another, with specified [TfLiteType].
   ///
   /// Throws [ArgumentError.notNull] if [buffer] is null.
-  static TensorBuffer createFrom(TensorBuffer buffer, TfLiteType dataType) {
+  static TensorBuffer createFrom(TensorBuffer buffer, int dataType) {
     SupportPreconditions.checkNotNull(buffer,
         message: "Cannot create a buffer from null");
     TensorBuffer result;
@@ -95,8 +95,8 @@ abstract class TensorBuffer {
     // The only scenario we need float array is FLOAT32->FLOAT32, or we can always use INT as
     // intermediate container.
     // The assumption is not true when we support other data types.
-    if (buffer.getDataType() == TfLiteType.float32 &&
-        dataType == TfLiteType.float32) {
+    if (buffer.getDataType() == TfLiteType.kTfLiteFloat32 &&
+        dataType == TfLiteType.kTfLiteFloat32) {
       List<double> data = buffer.getDoubleList();
       result.loadList(data, shape: buffer.shape);
     } else {
@@ -119,7 +119,7 @@ abstract class TensorBuffer {
   List<int> getShape() => shape;
 
   /// Returns the data type of this buffer.
-  TfLiteType getDataType();
+  int getDataType();
 
   /// Returns a List<double> of the values stored in this buffer. If the buffer is of different types
   /// than double, the values will be converted into double. For example, values in
